refactor: 공통 클래스 모듈 이동
