+++ maeil-mail/src/main/java/maeilmail/admin/AdminNotice.java
@@ -10,7 +10,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import maeilmail.support.BaseEntity;
+import maeilsupport.BaseEntity;
 
 @Entity
 @Getter
