+++ maeil-mail/src/main/java/maeilmail/bulksend/schedule/SendWeeklyQuestionScheduler.java
@@ -19,8 +19,8 @@ import maeilmail.question.QuestionSummary;
 import maeilmail.subscribe.command.domain.Subscribe;
 import maeilmail.subscribe.command.domain.SubscribeFrequency;
 import maeilmail.subscribe.command.domain.SubscribeRepository;
-import maeilmail.support.DateUtils;
 import maeilmail.support.DistributedSupport;
+import maeilsupport.DateUtils;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
