+++ maeil-mail/src/main/java/maeilmail/mail/MailEvent.java
@@ -9,7 +9,7 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import maeilmail.support.BaseEntity;
+import maeilsupport.BaseEntity;
 
 @Getter
 @Entity
