+++ maeil-mail/src/main/java/maeilmail/question/Question.java
@@ -13,7 +13,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import maeilmail.support.BaseEntity;
+import maeilsupport.BaseEntity;
 
 @Entity
 @Getter
