+++ maeil-mail/src/main/java/maeilmail/subscribe/command/domain/Subscribe.java
@@ -14,7 +14,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import maeilmail.question.QuestionCategory;
-import maeilmail.support.BaseEntity;
+import maeilsupport.BaseEntity;
 
 @Entity
 @Getter
