+++ maeil-mail/src/main/java/maeilmail/subscribe/command/domain/SubscribeQuestion.java
@@ -11,7 +11,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import maeilmail.question.Question;
-import maeilmail.support.BaseEntity;
+import maeilsupport.BaseEntity;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
