+++ maeil-mail/src/main/java/maeilmail/subscribe/command/domain/TemporalSubscribe.java
@@ -9,7 +9,7 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import maeilmail.support.BaseEntity;
+import maeilsupport.BaseEntity;
 
 @Entity
 @Getter
