+++ maeil-mail/src/main/java/maeilmail/subscribe/query/SubscribeQuestionQueryService.java
@@ -11,8 +11,8 @@ import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import maeilmail.question.QuestionCategory;
-import maeilmail.support.DateUtils;
-import maeilmail.support.PaginationResponse;
+import maeilsupport.DateUtils;
+import maeilsupport.PaginationResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.support.PageableExecutionUtils;
