+++ maeil-mail/src/test/java/maeilmail/question/QuestionApiTest.java
@@ -12,7 +12,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import java.util.Collections;
 import maeilmail.support.ApiTestSupport;
-import maeilmail.support.PaginationResponse;
+import maeilsupport.PaginationResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
