+++ maeil-mail/src/test/java/maeilmail/question/QuestionQueryServiceTest.java
@@ -9,7 +9,7 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.NoSuchElementException;
 import maeilmail.support.IntegrationTestSupport;
-import maeilmail.support.PaginationResponse;
+import maeilsupport.PaginationResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
